requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
